tests/e2e/pyproject.toml
[project]
name = "kiseki-e2e"
version = "0.0.0"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "protobuf>=5.0",
    "pytest>=8.0",
    "tenacity>=8.0",
    "pydantic>=2.0",
    "structlog>=24.0",
]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-ra", "--strict-markers", "--strict-config"]
markers = [
    "e2e: end-to-end tests against real server subprocess",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]

[tool.mypy]
disallow_untyped_defs = true
check_untyped_defs = true
